+++ category_encoders/__init__.py
@@ -22,8 +22,9 @@ from category_encoders.woe import WOEEncoder
 from category_encoders.m_estimate import MEstimateEncoder
 from category_encoders.james_stein import JamesSteinEncoder
 from category_encoders.cat_boost import CatBoostEncoder
+from category_encoders.mixed import MixedEncoder
 
-__version__ = '2.1.0'
+__version__ = '2.2.0'
 
 __author__ = 'willmcginnis'
 
@@ -42,5 +43,6 @@ __all__ = [
     'WOEEncoder',
     'MEstimateEncoder',
     'JamesSteinEncoder',
-    'CatBoostEncoder'
+    'CatBoostEncoder',
+    'MixedEncoder'
 ]